requirements.txt
docling>=2.15.1
pdfplumber>=0.11.4
PyMuPDF>=1.24.11
pytesseract>=0.3.10
easyocr>=1.7.2
python-docx>=1.1.2
beautifulsoup4>=4.12.3
langdetect>=1.0.9
Pillow>=10.4.0
